enable buildFuture: allow the website to include upcoming events
The website does not include events of which the date is after the current date,
e.g., a DiverSE Coffee talk will appear on the website only if the commit is
pushed a day after the coffee date.
Enabling this option would allow us to share the abstract before or during the
coffee and not context-switch and forget to do it the day after.